requirements.txt
PyVirtualDisplay==3.0
pyzmq==25.1.1
Requests==2.31.0
selenium==4.15.2
tabulate==0.9.0
